UDF
Support has been added for all currently supported boundary conditions for UDFs. This is enabled thorugh the "enable_udf" decorator function contained in fluent.py. This has not been thoroughly tested and may produced unexpected results.

The UDF class in fluent.py allows specification of the UDF

the base boundary condition class FluentBoundaryCondition has been given the methods "add_udf()" and "remove_udf()" to add and remove udf's. Refer to this class for documentation.

package io was renamed to fluentio due to interference with the native io package in python.

the module "engines.py" in the main folder has been deleted and it's contents moved to tui/fluent.py

some improvents to the parsing in classes have been implemented - specifically the direction vector in FluentFluidBoundaryCondition